Settings to set the editor theme explicitly
Three values are available now in the settings.
Auto means we read the theme kind from the vscode theme and sync that with the editor theme.
Dark means we always render the editor in the dark theme.
Light means we always render the editor in the light theme.
No support for high contract themes yet as the editor doesn't have that concept.